Refactor example_app, and fix associated tests
