--- tests/example_app/__init__.py
@@ -1,6 +1,0 @@
-import pale
-from tests.example_app import endpoints, resources
-
-_module_type = pale.ImplementationModule
-
-__all__ = ['endpoints', 'resources', '_module_type']
