+++ tests/example_app/api/__init__.py
@@ -1,0 +1,6 @@
+import pale
+from tests.example_app.api import endpoints, resources
+
+_module_type = pale.ImplementationModule
+
+__all__ = ['endpoints', 'resources', '_module_type']
