+++ tests/example_app/flask_app.py
@@ -7,7 +7,7 @@ from tests.example_app import api
 
 def create_pale_flask_app():
     """Creates a flask app, and registers a blueprint bound to pale."""
-    blueprint = flask.Blueprint('api', 'test.flask_example.api_blueprint')
+    blueprint = flask.Blueprint('api', 'tests.example_app')
     pale_flask_adapter.bind_blueprint(api, blueprint)
 
     app = flask.Flask(__name__)
